+++ backend/src/game/Dealer.ts
@@ -1,3 +1,4 @@
+import { GameService } from '../game-service';
 import { Game } from './Game';
 import { PokerTable } from './PokerTable';
 
@@ -19,6 +20,11 @@ export class Dealer {
     } else {
       pokerTable.getGame()?.getGameState().updateSeatToAct(pokerTable.getDealerPosition());
     }
+
+    GameService.eventEmitter.emit('startGame', pokerTable);
+
+    Dealer.dealCards(pokerTable);
+    Dealer.startTurn(pokerTable);
   }
 
   public static dealCards(pokerTable: PokerTable) {
@@ -35,7 +41,26 @@ export class Dealer {
       const player = seat.getPlayer();
       if (player) {
         player.setCards(deck.draw(2));
+        GameService.eventEmitter.emit('sendHoleCards', player);
       }
     });
+  }
+
+  public static startTurn(pokerTable: PokerTable) {
+    const game = pokerTable.getGame();
+    if (!game) {
+      throw new Error('Game not found');
+    }
+
+    const seatToAct = game.getGameState().getSeatToAct();
+    const seats = pokerTable.getSeats();
+
+    const seat = seats.find((seat) => seat.getSeatNumber() === seatToAct);
+
+    if (!seat) {
+      throw new Error(`Seat not found: ${seatToAct}`);
+    }
+
+    GameService.eventEmitter.emit('notifyPlayerToAct', pokerTable.getName(), seat.getSeatNumber());
   }
 }
