+++ backend/src/game/PokerTable.ts
@@ -5,6 +5,7 @@ import { SeatNotFoundError } from '../handlers/poker-tables/errors/SeatNotFoundE
 import { PlayerAlreadySeatedError } from '../handlers/poker-tables/errors/gen/PlayerAlreadySeatedError';
 import { PlayerNotFoundAtPokerTableError } from '../handlers/poker-tables/errors/gen/PlayerNotFoundAtPokerTableError';
 import { SeatTakenError } from '../handlers/poker-tables/errors/gen/SeatTakenError';
+import { Dealer } from './Dealer';
 import { Game } from './Game';
 import { Player } from './Player';
 import { Seat } from './Seat';
@@ -54,6 +55,9 @@ export class PokerTable {
 
           GameService.eventEmitter.emit('playerJoined', this, player, seatNumber);
 
+          if (this.isPokerTableReady()) {
+            Dealer.newGame(this);
+          }
           return Result.success();
         }
       }
